// src/modules/masters/presentation/http/validation/schemas/update-master-schedule-exception-payload.types.ts
import type { EMasterScheduleExceptionKind } from 'src/modules/masters/domain/entities/master-schedule-exception';

export interface IUpdateMasterScheduleExceptionPayload {
  startsAt?: string;
  endsAt?: string;
  kind?: EMasterScheduleExceptionKind;
  customStartTime?: string | null;
  customEndTime?: string | null;
  title?: string | null;
  note?: string | null;
}